apps/backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
saytex==0.1.6
gunicorn==23.0.0
python-dotenv==1.0.1
opensearch-py
sentence-transformers
transformers
torch
scikit-learn